separate getTermsEntriesStatus
